+++ app/services/ingest_csv.rb
@@ -1,7 +1,15 @@
 require "csv"
 
 class IngestCsv
-  ACCEPTABLE_INVALID_ATTRIBUTES = [:gdi, :intended_beneficiaries].freeze
+  ACCEPTABLE_INVALID_ATTRIBUTES = [
+    :gdi,
+    :intended_beneficiaries,
+    :report,
+    :receiving_organisation_name,
+    :receiving_organisation_type,
+    :date,
+    :collaboration_type,
+  ].freeze
 
   attr_accessor :csv, :filename
 
@@ -22,25 +30,33 @@ class IngestCsv
           next
         end
 
+        if activity == :missing_parent
+          write_log("Couldn't find parent Activity where roda_identifier_compound is #{row["parent_roda_identifier_compound"]}")
+          next
+        end
+
         Rails.logger.tagged(["IngestCsv", "activity:#{activity.id}"]) do
           attributes = IngestCsvRow.new(row).call
 
           # Ignore attributes that have been set to :skip
           attributes.delete_if { |_key, value| value == :skip }
 
-          activity.assign_attributes(attributes)
+          # Ignore parent_roda_identifier_compound attribute as there's no equivalent in RODA
+          attributes.delete("parent_roda_identifier_compound")
 
-          if activity.valid?
-            # Clean save
-            activity.save!
-          elsif (activity.errors.keys - ACCEPTABLE_INVALID_ATTRIBUTES).empty?
-            # Force validation if the only invalid fields are ones we're ok with
-            activity.save!(validate: false)
+          if has_transactions?
+            transaction = activity.transactions.find_or_initialize_by(
+              description: attributes["description"]
+            ) { |transaction|
+              transaction.ingested = true
+              transaction.report = Report.find_by!(financial_quarter: nil, fund: activity.associated_fund, organisation: activity.organisation)
+            }
+
+            attributes.delete("delivery_partner_identifier")
+
+            save_changes!(transaction, attributes)
           else
-            # Skip row and write message
-            write_log("Skipping Activity #{row}: #{activity.errors.full_messages}")
-            write_log("attributes: #{attributes.inspect}")
-            # binding.pry
+            save_changes!(activity, attributes)
           end
         end
       end
@@ -49,13 +65,73 @@ class IngestCsv
 
   private
 
+  def save_changes!(model, attributes)
+    model.assign_attributes(attributes)
+
+    if model.valid?
+      # Clean save
+      model.save!
+    elsif (model.errors.keys - ACCEPTABLE_INVALID_ATTRIBUTES).empty?
+      # Force validation if the only invalid fields are ones we're ok with
+      model.save!(validate: false)
+    else
+      # Skip row and log validation errors, ignoring the ones we've deemed acceptable
+      ACCEPTABLE_INVALID_ATTRIBUTES.each { |attr| model.errors.delete(attr) }
+
+      write_log("Skipping #{model.model_name.name} #{model.id}: #{model.errors.full_messages}")
+
+      attributes.sort.each do |k, v|
+        write_log("attributes: #{k.ljust(30, " ").slice(0...30)} => #{v}")
+      end
+    end
+  end
+
   def activity_for(row)
-    Activity.where.not(level: nil).find_by(
-      delivery_partner_identifier: row["delivery_partner_identifier"]
-    )
+    if has_parent_identifier?
+      parent = Activity.find_by(roda_identifier_compound: row["parent_roda_identifier_compound"])
+      return :missing_parent if parent.nil?
+
+      activity = Activity.find_or_initialize_by(
+        delivery_partner_identifier: row["delivery_partner_identifier"],
+        parent: parent,
+      ) { |activity|
+        activity.ingested = true
+        activity.funding_organisation_name = "Department for Business, Energy and Industrial Strategy"
+        activity.funding_organisation_reference = "GB-GOV-13"
+        activity.funding_organisation_type = "10"
+        activity.accountable_organisation_name = "Department for Business, Energy and Industrial Strategy"
+        activity.accountable_organisation_reference = "GB-GOV-13"
+        activity.accountable_organisation_type = "10"
+        activity.reporting_organisation = beis_organisation
+        activity.organisation = parent.organisation
+        activity.extending_organisation = parent.extending_organisation
+      }
+
+      # Always update these
+      activity.level = parent.child_level
+      activity.form_state = :complete
+
+      activity
+    else
+      Activity.where.not(level: nil).find_by(
+        delivery_partner_identifier: row["delivery_partner_identifier"]
+      )
+    end
+  end
+
+  def has_parent_identifier?
+    @has_parent_identifier ||= csv.first.headers.include?("parent_roda_identifier_compound")
+  end
+
+  def has_transactions?
+    @has_transactions ||= csv.first.headers.include?("transaction_type")
   end
 
   def write_log(message)
     Rails.logger.info(message)
+  end
+
+  def beis_organisation
+    @beis_organisation ||= Organisation.find_by(service_owner: true)
   end
 end